Refactor generate_default_trace into a library
This patch refactors the primary functionality in generate_default_trace
into a library and leaves the binary as a thin wrapper around the
library. This makes some code simpler as we really shoud not be
importing a binary into some other modules.
